Global styles engine: relocate resolveStyle and variation resolver
Rename `resolveStyles` to the singular `resolveStyle` and move the pure
cascade resolver from `block-editor` into `@wordpress/global-styles-engine`,
exposed via the engine's private API. The `useResolvedStyle` hook stays in
`block-editor` and owns all store access.

Also relocate `getVariationStylesWithRefValues` into the engine so
`resolveStyle` resolves the applied block style variation internally from its
slug. This collapses the earlier interim signature (a separate pre-resolved
`variationStyles` argument plus a slug used only for memoization) into a single
`ownVariation` argument, keeping the engine self-contained. `block-editor`
consumers (`block-style-variation`, `switch-section-style`) now unlock the
helper from the engine.

Relocated modules are authored in TypeScript to match engine conventions, and
their unit tests move alongside them.
