+++ packages/block-editor/src/hooks/block-style-variation.js
@@ -4,16 +4,23 @@
 import { getBlockTypes, store as blocksStore } from '@wordpress/blocks';
 import { useSelect } from '@wordpress/data';
 import { useMemo } from '@wordpress/element';
-import { toStyles, getBlockSelectors } from '@wordpress/global-styles-engine';
+import {
+	toStyles,
+	getBlockSelectors,
+	privateApis as globalStylesEnginePrivateApis,
+} from '@wordpress/global-styles-engine';
 
 /**
  * Internal dependencies
  */
 import { usePrivateStyleOverride } from './utils';
-import { getValueFromObjectPath } from '../utils/object';
 import { store as blockEditorStore } from '../store';
 import { globalStylesDataKey } from '../store/private-keys';
 import { unlock } from '../lock-unlock';
+
+const { getVariationStylesWithRefValues } = unlock(
+	globalStylesEnginePrivateApis
+);
 
 const VARIATION_PREFIX = 'is-style-';
 
@@ -173,77 +180,6 @@ export function BlockStyleVariationOverridesWithConfig( { config } ) {
 			) ) }
 		</>
 	);
-}
-
-/**
- * Retrieves any variation styles data and resolves any referenced values.
- *
- * @param {Object}    globalStyles A complete global styles object, containing settings and styles.
- * @param {string}    name         The name of the desired block type.
- * @param {variation} variation    The of the block style variation to retrieve data for.
- *
- * @return {Object|undefined} The global styles data for the specified variation.
- */
-export function getVariationStylesWithRefValues(
-	globalStyles,
-	name,
-	variation
-) {
-	if ( ! globalStyles?.styles?.blocks?.[ name ]?.variations?.[ variation ] ) {
-		return;
-	}
-
-	// Helper to recursively look for `ref` values to resolve.
-	const replaceRefs = ( variationStyles ) => {
-		Object.keys( variationStyles ).forEach( ( key ) => {
-			const value = variationStyles[ key ];
-
-			// Only process objects.
-			if ( typeof value === 'object' && value !== null ) {
-				// Process `ref` value if present.
-				if ( value.ref !== undefined ) {
-					if (
-						typeof value.ref !== 'string' ||
-						value.ref.trim() === ''
-					) {
-						// Remove invalid ref.
-						delete variationStyles[ key ];
-					} else {
-						// Resolve `ref` value.
-						const refValue = getValueFromObjectPath(
-							globalStyles,
-							value.ref
-						);
-
-						if ( refValue !== undefined && refValue !== null ) {
-							variationStyles[ key ] = refValue;
-						} else {
-							delete variationStyles[ key ];
-						}
-					}
-				} else {
-					// Recursively resolve `ref` values in nested objects.
-					replaceRefs( value );
-
-					// After recursion, if value is empty due to explicitly
-					// `undefined` ref value, remove it.
-					if ( Object.keys( value ).length === 0 ) {
-						delete variationStyles[ key ];
-					}
-				}
-			}
-		} );
-	};
-
-	// Deep clone variation node to avoid mutating it within global styles and losing refs.
-	const styles = JSON.parse(
-		JSON.stringify(
-			globalStyles.styles.blocks[ name ].variations[ variation ]
-		)
-	);
-	replaceRefs( styles );
-
-	return styles;
 }
 
 function useBlockStyleVariation( name, variation, clientId ) {
