+++ packages/global-styles-engine/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### New Features
+
+-   Added a private `resolveStyle` API that resolves the merged Global Styles cascade (root, block type, and applied block style variation) for a block, returning the inherited value and a per-leaf source map.
+-   Added a private `getVariationStylesWithRefValues` API that returns a block style variation's styles with any `{ ref }` envelopes resolved against the Global Styles tree.
+
 ## 1.18.0 (2026-07-14)
 
 ## 1.17.0 (2026-07-01)
