+++ packages/global-styles-engine/src/private-apis.ts
@@ -6,6 +6,8 @@ import {
 	getViewportBreakpoints,
 	getViewportBreakpointValueInPixels,
 } from './utils/viewport';
+import { resolveStyle } from './resolve-style';
+import { getVariationStylesWithRefValues } from './variation';
 import { lock } from './lock-unlock';
 
 export const privateApis = {};
@@ -13,4 +15,6 @@ lock( privateApis, {
 	getResponsiveMediaQueries,
 	getViewportBreakpoints,
 	getViewportBreakpointValueInPixels,
+	resolveStyle,
+	getVariationStylesWithRefValues,
 } );
