+++ packages/global-styles-engine/src/test/resolve-style.test.ts
@@ -1,20 +1,8 @@
 /**
- * External dependencies
- */
-import { render, screen } from '@testing-library/react';
-
-/**
- * WordPress dependencies
- */
-import { useSelect } from '@wordpress/data';
-
-/**
  * Internal dependencies
  */
-import { resolveStyles, privateHelpers } from '../resolve-styles';
-import { useResolvedStyles } from '../inherited-value-context';
+import { resolveStyle, privateHelpers } from '../resolve-style';
 
-import { globalStylesDataKey } from '../../../store/private-keys';
 const {
 	isExplicitEmpty,
 	isRefObject,
@@ -22,63 +10,7 @@ const {
 	deepMergeDroppingEmpties,
 } = privateHelpers;
 
-jest.mock( '@wordpress/data', () => ( {
-	useSelect: jest.fn(),
-	useDispatch: jest.fn( () => ( {} ) ),
-	useRegistry: jest.fn( () => ( {} ) ),
-	createSelector: jest.fn( ( callback ) => callback ),
-	createReduxStore: jest.fn(),
-	createRegistry: jest.fn(),
-	register: jest.fn(),
-	select: jest.fn(),
-	dispatch: jest.fn(),
-	combineReducers: jest.fn( ( reducers ) => reducers ),
-	subscribe: jest.fn(),
-	RegistryProvider: ( { children } ) => children,
-	RegistryConsumer: ( { children } ) => children( {} ),
-	AsyncModeProvider: ( { children } ) => children,
-	useRegistrySelect: jest.fn(),
-	useRegistryDispatch: jest.fn( () => ( {} ) ),
-	withSelect: ( mapStateToProps ) => ( Component ) => ( props ) =>
-		Component( {
-			...props,
-			...( mapStateToProps?.( () => ( {} ) ) || {} ),
-		} ),
-	withDispatch: () => ( Component ) => ( props ) => Component( props ),
-	withRegistry: ( Component ) => Component,
-} ) );
-
-jest.mock( '../../../store', () => ( {
-	store: { name: 'core/block-editor' },
-} ) );
-
-// `inherited-value-context.js` imports `store as blocksStore` from
-// `@wordpress/blocks` for `useOwnVariation`. The blocks store's transitive
-// import chain fails under this file's `@wordpress/data` mock (missing
-// `createSelector`), so stub the blocks module with just the shape needed.
-jest.mock( '@wordpress/blocks', () => ( {
-	store: { name: 'core/blocks' },
-	getBlockType: ( blockName ) =>
-		( {
-			'core/group': { title: 'Group' },
-			'core/heading': { title: 'Heading' },
-			'core/paragraph': { title: 'Paragraph' },
-		} )[ blockName ],
-} ) );
-
-// Short-circuit the variation-ref resolution path; ref handling is covered
-// by dedicated builder tests.
-jest.mock( '../../../hooks/block-style-variation', () => ( {
-	getVariationStylesWithRefValues: ( gs, blockName, variation ) =>
-		gs?.styles?.blocks?.[ blockName ]?.variations?.[ variation ] ?? null,
-	// `useResolvedStyles` derives the applied variation from the block's
-	// `className`; map `is-style-<slug>` to `<slug>` for these tests.
-	getVariationNameFromClass: ( className ) => {
-		const match = /is-style-([\w-]+)/.exec( className || '' );
-		return match ? match[ 1 ] : null;
-	},
-} ) );
-describe( 'resolveStyles – merged output', () => {
+describe( 'resolveStyle – merged output', () => {
 	describe( 'internals', () => {
 		test( 'isExplicitEmpty drops "", null, {} only', () => {
 			expect( isExplicitEmpty( '' ) ).toBe( true );
@@ -250,7 +182,7 @@ describe( 'resolveStyles – merged output', () => {
 		};
 
 		test( 'root only (layer 1)', () => {
-			const { value: out } = resolveStyles( {
+			const { value: out } = resolveStyle( {
 				blockName: 'core/paragraph',
 				globalStyles: gs,
 			} );
@@ -259,7 +191,7 @@ describe( 'resolveStyles – merged output', () => {
 		} );
 
 		test( 'block-default (layer 3) overrides element + root', () => {
-			const { value: out } = resolveStyles( {
+			const { value: out } = resolveStyle( {
 				blockName: 'core/heading',
 				globalStyles: gs,
 			} );
@@ -268,7 +200,7 @@ describe( 'resolveStyles – merged output', () => {
 		} );
 
 		test( 'own-variation (layer 4b) wins', () => {
-			const { value: out } = resolveStyles( {
+			const { value: out } = resolveStyle( {
 				blockName: 'core/heading',
 				ownVariation: 'plain',
 				globalStyles: gs,
@@ -299,7 +231,7 @@ describe( 'resolveStyles – merged output', () => {
 		};
 
 		test( 'no selectedState behaves as the default state (base value)', () => {
-			const { value: out } = resolveStyles( {
+			const { value: out } = resolveStyle( {
 				blockName: 'core/button',
 				globalStyles: gs,
 			} );
@@ -307,7 +239,7 @@ describe( 'resolveStyles – merged output', () => {
 		} );
 
 		test( 'explicit default selectedState is identical to base', () => {
-			const { value: out } = resolveStyles( {
+			const { value: out } = resolveStyle( {
 				blockName: 'core/button',
 				globalStyles: gs,
 				selectedState: DEFAULT_STATE,
@@ -316,7 +248,7 @@ describe( 'resolveStyles – merged output', () => {
 		} );
 
 		test( 'pseudo state layers the block `:hover` slice over base', () => {
-			const { value: out } = resolveStyles( {
+			const { value: out } = resolveStyle( {
 				blockName: 'core/button',
 				globalStyles: gs,
 				selectedState: HOVER_STATE,
@@ -326,7 +258,7 @@ describe( 'resolveStyles – merged output', () => {
 		} );
 
 		test( 'base-only leaves still inherit under a selected state (CSS cascade)', () => {
-			const { value: out } = resolveStyles( {
+			const { value: out } = resolveStyle( {
 				blockName: 'core/button',
 				globalStyles: gs,
 				selectedState: HOVER_STATE,
@@ -336,7 +268,7 @@ describe( 'resolveStyles – merged output', () => {
 		} );
 
 		test( 'responsive state with no Global Styles slice falls back to base', () => {
-			const { value: out } = resolveStyles( {
+			const { value: out } = resolveStyle( {
 				blockName: 'core/button',
 				globalStyles: gs,
 				selectedState: MOBILE_STATE,
@@ -361,7 +293,7 @@ describe( 'resolveStyles – merged output', () => {
 					},
 				},
 			};
-			const { value: out } = resolveStyles( {
+			const { value: out } = resolveStyle( {
 				blockName: 'core/button',
 				globalStyles: responsiveGs,
 				selectedState: MOBILE_STATE,
@@ -374,7 +306,7 @@ describe( 'resolveStyles – merged output', () => {
 		} );
 
 		test( 'source map attributes a state-won leaf to its originating layer', () => {
-			const { value, sources } = resolveStyles( {
+			const { value, sources } = resolveStyle( {
 				blockName: 'core/button',
 				globalStyles: gs,
 				selectedState: HOVER_STATE,
@@ -384,12 +316,12 @@ describe( 'resolveStyles – merged output', () => {
 		} );
 
 		test( 'memoized variant keys distinct states separately', () => {
-			const { value: base } = resolveStyles( {
+			const { value: base } = resolveStyle( {
 				blockName: 'core/button',
 				globalStyles: gs,
 				selectedState: DEFAULT_STATE,
 			} );
-			const { value: hover } = resolveStyles( {
+			const { value: hover } = resolveStyle( {
 				blockName: 'core/button',
 				globalStyles: gs,
 				selectedState: HOVER_STATE,
@@ -414,7 +346,7 @@ describe( 'resolveStyles – merged output', () => {
 					},
 				},
 			};
-			const { value, sources } = resolveStyles( {
+			const { value, sources } = resolveStyle( {
 				blockName: 'core/button',
 				globalStyles: gs,
 			} );
@@ -442,7 +374,7 @@ describe( 'resolveStyles – merged output', () => {
 					},
 				},
 			};
-			const { value, sources } = resolveStyles( {
+			const { value, sources } = resolveStyle( {
 				blockName: 'core/button',
 				globalStyles: gs,
 			} );
@@ -458,7 +390,7 @@ describe( 'resolveStyles – merged output', () => {
 					},
 				},
 			};
-			const { value } = resolveStyles( {
+			const { value } = resolveStyle( {
 				blockName: 'core/heading',
 				globalStyles: gs,
 			} );
@@ -473,7 +405,7 @@ describe( 'resolveStyles – merged output', () => {
 					},
 				},
 			};
-			const { value } = resolveStyles( {
+			const { value } = resolveStyle( {
 				blockName: 'core/paragraph',
 				globalStyles: gs,
 			} );
@@ -491,7 +423,7 @@ describe( 'resolveStyles – merged output', () => {
 					},
 				},
 			};
-			const { value } = resolveStyles( {
+			const { value } = resolveStyle( {
 				blockName: 'core/button',
 				globalStyles: gs,
 				selectedState: HOVER_STATE,
@@ -512,7 +444,7 @@ describe( 'resolveStyles – merged output', () => {
 					},
 				},
 			};
-			const { value: out } = resolveStyles( {
+			const { value: out } = resolveStyle( {
 				blockName: 'core/heading',
 				globalStyles: gs,
 			} );
@@ -529,7 +461,7 @@ describe( 'resolveStyles – merged output', () => {
 					},
 				},
 			};
-			const { value: out } = resolveStyles( {
+			const { value: out } = resolveStyle( {
 				blockName: 'core/group',
 				globalStyles: gs,
 			} );
@@ -540,13 +472,13 @@ describe( 'resolveStyles – merged output', () => {
 	describe( 'hydration + edge cases', () => {
 		test( 'falsy globalStyles returns {}', () => {
 			expect(
-				resolveStyles( {
+				resolveStyle( {
 					blockName: 'core/heading',
 					globalStyles: null,
 				} ).value
 			).toEqual( {} );
 			expect(
-				resolveStyles( {
+				resolveStyle( {
 					blockName: 'core/heading',
 					globalStyles: {},
 				} ).value
@@ -555,7 +487,7 @@ describe( 'resolveStyles – merged output', () => {
 
 		test( 'missing blockName returns {}', () => {
 			expect(
-				resolveStyles( {
+				resolveStyle( {
 					globalStyles: {
 						styles: { typography: { fontSize: '16px' } },
 					},
@@ -565,7 +497,7 @@ describe( 'resolveStyles – merged output', () => {
 
 		test( 'unknown block still inherits from root', () => {
 			const gs = { styles: { typography: { fontSize: '16px' } } };
-			const { value: out } = resolveStyles( {
+			const { value: out } = resolveStyle( {
 				blockName: 'core/does-not-exist',
 				globalStyles: gs,
 			} );
@@ -580,7 +512,7 @@ describe( 'resolveStyles – merged output', () => {
 					color: { text: 'var:preset|color|vivid-red' },
 				},
 			};
-			const { value: out } = resolveStyles( {
+			const { value: out } = resolveStyle( {
 				blockName: 'core/paragraph',
 				globalStyles: gs,
 			} );
@@ -593,7 +525,7 @@ describe( 'resolveStyles – merged output', () => {
 					color: { text: 'var(--wp--preset--color--vivid-red)' },
 				},
 			};
-			const { value: out } = resolveStyles( {
+			const { value: out } = resolveStyle( {
 				blockName: 'core/paragraph',
 				globalStyles: gs,
 			} );
@@ -614,7 +546,7 @@ describe( 'resolveStyles – merged output', () => {
 					css: ':root { --x: 1; }',
 				},
 			};
-			const { value: out } = resolveStyles( {
+			const { value: out } = resolveStyle( {
 				blockName: 'core/heading',
 				globalStyles: gs,
 			} );
@@ -631,7 +563,7 @@ describe( 'resolveStyles – merged output', () => {
 					},
 				},
 			};
-			const { value: out } = resolveStyles( {
+			const { value: out } = resolveStyle( {
 				blockName: 'core/paragraph',
 				globalStyles: gs,
 			} );
@@ -670,7 +602,7 @@ describe( 'resolveStyles – merged output', () => {
 		};
 
 		test( 'returns value and source map from the same merge', () => {
-			const { value, sources } = resolveStyles( {
+			const { value, sources } = resolveStyle( {
 				blockName: 'core/heading',
 				ownVariation: 'plain',
 				globalStyles: gs,
@@ -686,7 +618,7 @@ describe( 'resolveStyles – merged output', () => {
 		} );
 
 		test( 'records preserved element sub-tree source paths', () => {
-			const { sources } = resolveStyles( {
+			const { sources } = resolveStyle( {
 				blockName: 'core/paragraph',
 				globalStyles: gs,
 			} );
@@ -697,14 +629,14 @@ describe( 'resolveStyles – merged output', () => {
 	} );
 } );
 
-describe( 'resolveStyles – memoization', () => {
+describe( 'resolveStyle – memoization', () => {
 	test( 'returns the same inheritance object identity for identical keys', () => {
 		const gs = { styles: { typography: { fontSize: '16px' } } };
-		const a = resolveStyles( {
+		const a = resolveStyle( {
 			blockName: 'core/paragraph',
 			globalStyles: gs,
 		} );
-		const b = resolveStyles( {
+		const b = resolveStyle( {
 			blockName: 'core/paragraph',
 			globalStyles: gs,
 		} );
@@ -722,11 +654,11 @@ describe( 'resolveStyles – memoization', () => {
 				},
 			},
 		};
-		const { value: a } = resolveStyles( {
+		const { value: a } = resolveStyle( {
 			blockName: 'core/paragraph',
 			globalStyles: gs,
 		} );
-		const { value: b } = resolveStyles( {
+		const { value: b } = resolveStyle( {
 			blockName: 'core/heading',
 			globalStyles: gs,
 		} );
@@ -738,11 +670,11 @@ describe( 'resolveStyles – memoization', () => {
 	test( 'different globalStyles reference → re-computed', () => {
 		const gs1 = { styles: { typography: { fontSize: '16px' } } };
 		const gs2 = { styles: { typography: { fontSize: '18px' } } };
-		const { value: a } = resolveStyles( {
+		const { value: a } = resolveStyle( {
 			blockName: 'core/paragraph',
 			globalStyles: gs1,
 		} );
-		const { value: b } = resolveStyles( {
+		const { value: b } = resolveStyle( {
 			blockName: 'core/paragraph',
 			globalStyles: gs2,
 		} );
@@ -752,7 +684,7 @@ describe( 'resolveStyles – memoization', () => {
 	} );
 
 	test( 'falsy globalStyles delegates to the pure builder', () => {
-		const { value: a } = resolveStyles( {
+		const { value: a } = resolveStyle( {
 			blockName: 'core/paragraph',
 			globalStyles: null,
 		} );
@@ -760,243 +692,11 @@ describe( 'resolveStyles – memoization', () => {
 	} );
 } );
 
-describe( 'useResolvedStyles hook', () => {
-	beforeEach( () => {
-		useSelect.mockReset();
-	} );
-
-	function Probe( { blockName, className, selectedState } ) {
-		const v = useResolvedStyles( blockName, className, selectedState );
-		return <div data-testid="probe">{ JSON.stringify( v ) }</div>;
-	}
-
-	// The hook issues two `useSelect` reads: the merged Global Styles
-	// payload (block-editor store) and the block's registered styles
-	// (blocks store). Feed both from one stub.
-	function mockStores( rawGlobalStyles, blockStyles = [] ) {
-		useSelect.mockImplementation( ( mapSelect ) =>
-			mapSelect( () => ( {
-				// `globalStylesDataKey` holds the BARE merged styles tree
-				// in production (see `editor/src/components/provider/
-				// use-block-editor-settings.js:237`,
-				// `mergedGlobalStyles.styles`), not the wrapped
-				// `{ settings, styles }` envelope. The hook wraps the bare
-				// tree before passing to the builder.
-				getSettings: () => ( {
-					[ globalStylesDataKey ]: rawGlobalStyles,
-				} ),
-				getBlockStyles: () => blockStyles,
-			} ) )
-		);
-	}
-
-	test( 'returns empty value and sources when no block name is given', () => {
-		mockStores( { typography: { fontSize: '16px' } } );
-		render( <Probe /> );
-		expect( screen.getByTestId( 'probe' ) ).toHaveTextContent(
-			'{"value":{},"sources":{}}'
-		);
-	} );
-
-	test( 'returns empty value and sources during hydration', () => {
-		mockStores( null );
-		render( <Probe blockName="core/heading" /> );
-		expect( screen.getByTestId( 'probe' ) ).toHaveTextContent(
-			'{"value":{},"sources":{}}'
-		);
-	} );
-
-	test( 'reads the block element passthrough from the merged payload', () => {
-		mockStores( {
-			typography: { fontSize: '16px' },
-			elements: { h2: { typography: { fontSize: '24px' } } },
-		} );
-		render( <Probe blockName="core/heading" /> );
-		const parsed = JSON.parse( screen.getByTestId( 'probe' ).textContent );
-		// The h2 element styles are preserved under the nested
-		// `elements` passthrough, not folded up to the top level.
-		expect( parsed.value.elements.h2.typography.fontSize ).toBe( '24px' );
-		expect(
-			parsed.sources[ 'elements.h2.typography.fontSize' ].layer
-		).toBe( 'root' );
-	} );
-} );
-
-/**
- * Regression suite for the production data-shape contract.
- *
- * `settings[ globalStylesDataKey ]` carries the BARE merged Global
- * Styles tree — `{ typography: {...}, color: {...}, blocks: {...},
- * elements: {...}, ... }` — produced by
- * `editor/src/components/provider/use-block-editor-settings.js:237`
- * (`mergedGlobalStyles.styles`). It is NOT the wrapped
- * `{ settings, styles }` envelope.
- *
- * Pre-hot-fix, the hook passed the bare tree directly to
- * `resolveStyles`, which destructures `const { styles } =
- * globalStyles` and so saw `undefined` and early-returned `{}`. Every
- * panel got an empty `inheritedValue`; nothing surfaced in the
- * inspector. None of the original suites caught it because their
- * fixtures all set `[ globalStylesDataKey ]: { styles: { ... } }` —
- * encoding the wrong shape assumption.
- *
- * This suite uses fixtures that mirror the production producer's
- * output and asserts the hook → builder pipeline yields a
- * non-empty `inheritedValue` end-to-end. A future regression on the
- * data-shape contract — at the producer, the wrapping step inside the
- * hook, or the builder's destructure — surfaces here in CI.
- */
-describe( 'useResolvedStyles – production bare-tree shape', () => {
-	beforeEach( () => {
-		useSelect.mockReset();
-	} );
-
-	function Probe( { blockName, className } ) {
-		const { value } = useResolvedStyles( blockName, className );
-		return <div data-testid="probe">{ JSON.stringify( value ) }</div>;
-	}
-
-	/**
-	 * Representative bare-tree fixture shaped like a real
-	 * `mergedGlobalStyles.styles` payload from a child theme that:
-	 *
-	 *   - Sets a root text color, body background, and root line height.
-	 *   - Overrides `core/heading` with a font-family + weight.
-	 *   - Overrides the `h2` element specifically with a smaller
-	 *     font size and a different color.
-	 *   - Registers a `plain` variation under `core/quote` that
-	 *     drops the border and recolors the text.
-	 *   - Includes a `{ ref: 'styles.color.background' }` envelope on
-	 *     the `core/group` background to verify ref resolution still
-	 *     works end-to-end (ref envelopes resolve against the wrapped
-	 *     `{ styles }` tree). This is block-sourced, so it survives the
-	 *     non-cascading root drop.
-	 */
-	const productionShapeFixture = {
-		typography: { lineHeight: '1.6' },
-		color: {
-			text: '#1a1a1a',
-			background: '#ffffff',
-		},
-		elements: {
-			link: { color: { text: '#0073aa' } },
-		},
-		blocks: {
-			'core/heading': {
-				typography: {
-					fontFamily: 'serif',
-					fontWeight: '600',
-				},
-				elements: {
-					h2: {
-						typography: { fontSize: '28px' },
-						color: { text: '#444444' },
-					},
-				},
-			},
-			'core/quote': {
-				border: { width: '4px', color: '#cccccc' },
-				variations: {
-					plain: {
-						border: { width: '0px' },
-						color: { text: '#666666' },
-					},
-				},
-			},
-			'core/group': {
-				color: { background: { ref: 'styles.color.background' } },
-			},
-		},
-	};
-
-	function mountWithFixture( ui, rawGlobalStyles = productionShapeFixture ) {
-		useSelect.mockImplementation( ( mapSelect ) =>
-			mapSelect( () => ( {
-				getSettings: () => ( {
-					[ globalStylesDataKey ]: rawGlobalStyles,
-				} ),
-				getBlockStyles: () => [],
-			} ) )
-		);
-		return render( ui );
-	}
-
-	test( 'root + block override merge yields non-empty payload at the panel boundary', () => {
-		mountWithFixture( <Probe blockName="core/heading" /> );
-		const parsed = JSON.parse( screen.getByTestId( 'probe' ).textContent );
-		expect( parsed ).toMatchObject( {
-			typography: {
-				lineHeight: '1.6', // from root layer
-				fontFamily: 'serif', // from block override
-				fontWeight: '600', // from block override
-			},
-			color: { text: '#1a1a1a' }, // from root layer
-		} );
-	} );
-
-	test( 'block element override is preserved under the elements passthrough', () => {
-		mountWithFixture( <Probe blockName="core/heading" /> );
-		const parsed = JSON.parse( screen.getByTestId( 'probe' ).textContent );
-		// Block-level h2 override stays nested under `elements.h2`.
-		expect( parsed.elements.h2.typography.fontSize ).toBe( '28px' );
-		expect( parsed.elements.h2.color.text ).toBe( '#444444' );
-		// Root + block overrides still merge at the top level.
-		expect( parsed.typography.lineHeight ).toBe( '1.6' );
-	} );
-
-	test( 'variation override layers on top of block override', () => {
-		mountWithFixture(
-			<Probe blockName="core/quote" className="is-style-plain" />
-		);
-		const parsed = JSON.parse( screen.getByTestId( 'probe' ).textContent );
-		expect( parsed.border.width ).toBe( '0px' ); // variation wins
-		expect( parsed.border.color ).toBe( '#cccccc' ); // block passthrough
-		expect( parsed.color.text ).toBe( '#666666' ); // variation wins
-	} );
-
-	test( '{ ref } envelope resolves against the bare tree at the leaf', () => {
-		mountWithFixture( <Probe blockName="core/group" /> );
-		const parsed = JSON.parse( screen.getByTestId( 'probe' ).textContent );
-		expect( parsed.color.background ).toBe( '#ffffff' );
-	} );
-
-	test( 'block with no overrides still inherits root + element layers', () => {
-		mountWithFixture( <Probe blockName="core/paragraph" /> );
-		const parsed = JSON.parse( screen.getByTestId( 'probe' ).textContent );
-		expect( parsed.typography.lineHeight ).toBe( '1.6' );
-		// The root link element styles stay nested under `elements.link`.
-		expect( parsed.elements.link.color.text ).toBe( '#0073aa' );
-	} );
-
-	test( 'panel-readable keys are absent when the producer accidentally double-wraps the data', () => {
-		// Regression guard against re-introducing the original bug. If
-		// a future refactor accidentally double-wraps the data —
-		// i.e. stores `{ styles: { typography: ... } }` at the dataKey
-		// instead of the bare tree — the Provider's wrapper produces
-		// `{ styles: { styles: { typography: ... } } }`, the builder
-		// destructures `styles = { styles: { typography: ... } }`, and
-		// `pickLayerRootContribution` treats `styles` as a leaf-bearing
-		// key so the merged payload is `{ styles: { typography: ... } }`
-		// — non-empty in object terms, but the panel-readable keys
-		// (`typography`, `color`, ...) are absent at the top level. The
-		// user observes the same thing they did before the hot fix: no
-		// inherited values surface in the inspector. This case asserts
-		// the panel-readable keys are missing, which is what the panels
-		// actually consume.
-		mountWithFixture( <Probe blockName="core/heading" />, {
-			styles: productionShapeFixture,
-		} );
-		const parsed = JSON.parse( screen.getByTestId( 'probe' ).textContent );
-		expect( parsed.typography ).toBeUndefined();
-		expect( parsed.color ).toBeUndefined();
-	} );
-} );
-
-describe( 'resolveStyles – non-cascading root drop', () => {
+describe( 'resolveStyle – non-cascading root drop', () => {
 	// Each call builds against a fresh `globalStyles` object so the
 	// identity-keyed memo never returns a cross-test cache hit.
 	const build = ( styles, extra = {} ) =>
-		resolveStyles( {
+		resolveStyle( {
 			blockName: 'core/paragraph',
 			globalStyles: { styles },
 			...extra,
@@ -1065,7 +765,7 @@ describe( 'resolveStyles – non-cascading root drop', () => {
 	} );
 
 	test( 'resolves a theme-file background image url via _links', () => {
-		const { value } = resolveStyles( {
+		const { value } = resolveStyle( {
 			blockName: 'core/paragraph',
 			globalStyles: {
 				styles: {
